#2: Schema update. Linking run elements to samples by embedding.
