+++ rest_api/aggregation/__init__.py
@@ -1,0 +1,2 @@
+__author__ = 'mwham'
+from . import database_side, server_side
